requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.3.0
altair==5.5.0
annotated-types==0.6.0
anyio==4.9.0
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
chevron==0.14.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
dataclasses-json==0.5.14
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
easydict==1.10
exceptiongroup==1.2.1
faiss-cpu==1.11.0
fastapi==0.110.2
fastapi-socketio==0.0.10
filelock==3.12.0
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.4
imageio==2.30.0
importlib-metadata==7.0.0
insightface==0.7.3
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.4
langchain==0.3.26
langchain-community==0.0.34
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.2
Lazify==0.4.0
lazy_loader==0.2
literalai==0.0.507
llama_cpp_python==0.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.1
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.26.4
ollama==0.1.8
onnx==1.14.0
onnxruntime==1.15.0
openai==1.23.6
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.1
packaging==23.2
pandas==2.3.0
Pillow==9.0.1
plotly==6.2.0
prettytable==3.7.0
protobuf==4.23.2
psutil==5.9.5
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.8.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==4.1.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.2
SQLAlchemy==2.0.29
sse-starlette==2.3.6
starlette==0.37.2
starlette-context==0.3.6
streamlit==1.46.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
threadpoolctl==3.1.0
tifffile==2023.4.12
tk==0.1.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2+cu118
tornado==6.5.1
tqdm==4.65.0
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uptrace==1.24.0
urllib3==2.0.2
uvicorn==0.25.0
watchdog==6.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
zstandard==0.23.0
